Revert "Trace Metadata: Convert kernel_stack vector to usize"
* Rather modify the FPT to use u64

This reverts commit 7d88575fabe2cc3f251061946dadcb65afb57ba9.

+++ vinter_rust/vinter_common/src/trace.rs
@@ -12,7 +12,7 @@ pub struct Metadata {
     /// currently in kernel mode?
     pub in_kernel: bool,
     /// kernel stack trace (frame pointer walk)
-    pub kernel_stacktrace: Vec<usize>,
+    pub kernel_stacktrace: Vec<u64>,
 }
 
 #[derive(Debug, Encode, Decode)]
